python-microservice/requirements.txt

fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
Pillow==10.0.1
pytesseract==0.3.10
pdfplumber==0.10.3
python-dotenv==1.0.0
